refactor: (#30) 모달 제어 로직 분리
