+++ src/pages/main/ui/LoginModal.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useId, useRef } from 'react';
+import { useId } from 'react';
 import { ArrowLeft, Eye, UsersRound } from 'lucide-react';
 import { useForm } from 'react-hook-form';
+import { useDialog } from '@/shared/lib/useDialog';
 
 interface LoginModalProps {
   isOpen: boolean;
@@ -13,7 +14,7 @@ interface LoginFormValues {
 }
 
 const LoginModal = ({ isOpen, onClose }: LoginModalProps) => {
-  const dialogRef = useRef<HTMLDialogElement>(null);
+  const dialogRef = useDialog({ isOpen, onClose });
   const emailId = useId();
   const passwordId = useId();
 
@@ -23,41 +24,6 @@ const LoginModal = ({ isOpen, onClose }: LoginModalProps) => {
       password: '',
     },
   });
-
-  useEffect(() => {
-    const dialogElement = dialogRef.current;
-
-    if (!dialogElement) {
-      return;
-    }
-
-    if (isOpen && !dialogElement.open) {
-      dialogElement.showModal();
-    }
-
-    if (!isOpen && dialogElement.open) {
-      dialogElement.close();
-    }
-  }, [isOpen]);
-
-  useEffect(() => {
-    const dialogElement = dialogRef.current;
-
-    if (!dialogElement) {
-      return;
-    }
-
-    const handleCancel = (event: Event) => {
-      event.preventDefault();
-      onClose();
-    };
-
-    dialogElement.addEventListener('cancel', handleCancel);
-
-    return () => {
-      dialogElement.removeEventListener('cancel', handleCancel);
-    };
-  }, [onClose]);
 
   const handleClose = () => {
     reset();
